producer/requirements.txt
kafka-python
dotenv
requests